requirements.txt
# Core
pyspark==3.5.0
delta-spark==2.4.0

# API extraction
requests>=2.31.0

# Kafka
confluent-kafka>=2.3.0

# Airflow (for local development — production uses Docker image)
apache-airflow==2.8.1
apache-airflow-providers-apache-spark==4.7.1
apache-airflow-providers-slack==8.6.0

# dbt
dbt-core==1.7.0
dbt-databricks==1.7.0

# Data quality
great-expectations==0.18.8

# ML
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Linting
flake8>=6.1.0
black>=23.12.0

# Utilities
python-dotenv>=1.0.0
